test: split backtesting and feature caching suites (#257)
* test: split backtesting and feature caching suites

* Merge branch 'develop' into work

* Merge branch 'develop' into work

# Conflicts:
#	tests/unit/database/conftest.py
